requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0

SQLAlchemy==2.0.36
alembic==1.14.0

pydantic==2.9.2
python-multipart==0.0.9

pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==11.0.0

openai==1.52.2
loguru==0.7.3
python-dotenv==1.0.1
